seedData/requirements.txt
Faker==9.8.3
mysql-connector-python==8.0.27